UPDATED METRONOME EXCEL FILE
Marc created a list of values that will determine the speed of servo for each BPM setting.

With them I was able to come up with a formula that can give us a more accurate value.

So far I have tested it with the RadioShack Micro servo, the SG90 is still to be tested.

For the SG90 use the column that says, MAX 3000 MIN 1000(speed). This will be included in the source code as a commented out array.
So you can just comment out the other array and use the other one.